+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added the query to get the organizations without a sales manager
+
 ## [0.17.1] - 2022-05-31
 
 ### Added